+++ core/src/main/java/es/urjc/etsii/grafo/util/ValidationUtil.java
@@ -1,6 +1,7 @@
 package es.urjc.etsii.grafo.util;
 
 import es.urjc.etsii.grafo.io.Instance;
+import es.urjc.etsii.grafo.solution.Move;
 import es.urjc.etsii.grafo.solution.Solution;
 
 import java.util.List;
@@ -42,6 +43,17 @@ public class ValidationUtil {
     public static <S extends Solution<S,I>, I extends Instance> void positiveTTB(S solution){
         if(solution.getLastModifiedTime() < 0){
             throw new AssertionError(String.format("Last modified time cannot be negative, current value: %s. Remember to call Solution::updateLastModifiedTime if you have modified the solution without using a Move!", solution.getLastModifiedTime()));
+        }
+    }
+
+    public static <S extends Solution<S,I>, I extends Instance> boolean scoreUpdate(S solution, double old, Move<S,I> move){
+        double newScore = solution.getScore();
+        double diff = newScore - old;
+        double expected = move.getValue();
+        if(DoubleComparator.equals(diff, expected)){
+            return true;
+        } else {
+            throw new AssertionError(String.format("Score change validation failed: %s - %s = %s != %s, current move: %s, solution state with move applied: %s", newScore, old, diff, expected, move, solution));
         }
     }
 }